adding folder structure and redux
