+++ src/reducers/rootReducer.js
@@ -1,0 +1,4 @@
+import { combineReducers } from 'redux';
+// import { holidays } from './holidays';
+
+export const rootReducer = combineReducers({  })
