+++ src/sagas/rootSaga.js
@@ -1,0 +1,6 @@
+import { all, fork } from 'redux-saga/effects';
+// import { holidays } from "./holidays";
+
+export function* rootSaga() {
+    yield all([]);
+}
